+++ src/django/api/views.py
@@ -3785,3 +3785,22 @@ class NonstandardFieldsViewSet(mixins.ListModelMixin,
             NonstandardField.DEFAULT_FIELDS.keys() | nonstandard_field_set)
 
         return Response(field_list)
+
+
+@api_view(['GET'])
+@permission_classes([IsAdminUser])
+def get_geocoding(request):
+    """
+    Returns geocoded name and address.
+    """
+    country_code = request.query_params.get('country_code', None)
+    address = request.query_params.get('address', None)
+
+    if country_code is None:
+        raise BadRequestException('Missing country code')
+
+    if address is None:
+        raise BadRequestException('Missing address')
+
+    geocode_result = geocode_address(address, country_code)
+    return Response(geocode_result)